refactor(migration): Isolate migration models into versioned packages
This change improves code organization and prevents package name collisions between different migration versions by ensuring that each migration's models are self-contained.

The following changes were made:
- Renamed the package in `migration/v6` from `models` to `v6`.
- Renamed the package in `migration/v7` from `models` to `v7`.
- Updated all associated test files and import paths to reflect the package structure.
