+++ migration/v6/model.go
@@ -1,4 +1,4 @@
-package models
+package v6
 
 import "time"
