+++ migration/v7/models_suite_test.go
@@ -1,4 +1,4 @@
-package models_test
+package v7_test
 
 import (
 	. "github.com/onsi/ginkgo/v2"
