+++ migration/v7/router_groups.go
@@ -1,4 +1,4 @@
-package models
+package v7
 
 import (
 	"errors"
